requirements.txt
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
wandb>=0.15.0
tqdm>=4.65.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
tensorboard>=2.13.0
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0
python-dotenv>=1.0.0
joblib>=1.2.0
pillow>=9.5.0